FIX: Remove broadcast_shape from the DimMax trait
While calling co_broadcast directly is less convenient, for now they are
two different functions.
